requirements.txt
channels
djangorestframework
django-filter
django-cors-headers
django-python3-ldap
asgi_redis
